requirements.txt
django>=3.1.2,<3.2
django-cors-headers>=3.5.0,<3.6
django-mysql>=3.9.0,<4
graphene>=2.1.8,<2.2
graphene-django>=2.13.0,<2.14
mysqlclient>=2.0.1,<2.1
# uWSGI>=2.0.19,<2.1
django-graphql-jwt
django-debug-toolbar